feat: migrate to @takazudo/zudo-doc 2.0.0 (Collapse Wiring Shells)
Bring this consuming docs site up to the zudo-doc 2.0.0 MAJOR release.

- deps: bump the zudo-doc group ^1.2.0 -> ^2.0.0 in lockstep
  (@takazudo/zudo-doc, @takazudo/zudo-doc-history-server, create-zudo-doc);
  the @takazudo/zfb* group is left untouched (0.1.0-next.69).
- host-wiring collapse: adopt the new pages/lib/_chrome.ts (single chrome
  adapter via createChrome) + _route-context.ts (createRouteContext), and
  remove ~24 now-obsolete per-component shells (_*-with-defaults,
  _doc-page-*, _*-nav, _doc-*-area, _compose-meta-title, _head-with-defaults,
  _math-block, route-enumerators, …) plus pages/_mdx-components.ts — their
  logic is now provided by createChrome + the package's createMdxComponents.
- repoint route pages (index, [locale]/index, docs/[[...slug]],
  [locale]/docs/[[...slug]]) to the _chrome adapter.
- _chrome mdxExtras: retain the zcss-specific CssPreview / TailwindPreview
  islands + showcase stubs (Badge/Dialog/MDXRemote/ProductCard/Teleport);
  img base-rewrite, the imageEnlarge <p> override, admonitions, CodeGroup,
  Tabs, MathBlock and locale-bound nav are now baked by the package.
- thin shims _doc-route-entries / _nav-source-docs / src/config/
  tag-vocabulary-types now re-export from _route-context / the package
  (matches create-zudo-doc@2.0.0 templates; template-drift clean).
- typography: move the Noto Sans JP webfont load from the removed custom
  host <head> into src/styles/global.css (2.0.0 collapsed the host head
  into package chrome, which has no custom-font hook) — applies uniformly
  to the homepage and every doc page via --font-noto.
- allowlist: drop the deleted _mdx-components / _head-with-defaults entries,
  add pages/lib/_chrome.ts, note the font move in global.css.

Gates: pin-parity, template-drift, check (tsc), build (240 pages),
generate:css-wisdom + setup-doc-skill — all pass.
