+++ package.json
@@ -28,8 +28,8 @@
     "@takazudo/zfb": "0.1.0-next.69",
     "@takazudo/zfb-runtime": "0.1.0-next.69",
     "@takazudo/zfb-adapter-cloudflare": "0.1.0-next.69",
-    "@takazudo/zudo-doc": "^1.2.0",
-    "@takazudo/zudo-doc-history-server": "^1.2.0",
+    "@takazudo/zudo-doc": "^2.0.0",
+    "@takazudo/zudo-doc-history-server": "^2.0.0",
     "zod": "^4.0.0",
     "preact": "^10.29.1",
     "preact-render-to-string": "^6.6.6",
@@ -50,7 +50,7 @@
     "typescript": "^5.9.0",
     "@types/node": "^22.0.0",
     "@types/react": "^19.2.0",
-    "create-zudo-doc": "^1.2.0",
+    "create-zudo-doc": "^2.0.0",
     "html-validate": "^10.0.0",
     "lefthook": "^1.11.0",
     "pagefind": "^1.4.0",
