--- pages/lib/_doc-body-end.tsx
@@ -1,8 +1,0 @@
-/** @jsxRuntime automatic */
-/** @jsxImportSource preact */
-// Host thin-stub — see @takazudo/zudo-doc/doc-body-end (epic #2344, S7).
-import { settings } from "@/config/settings";
-import { createDocBodyEnd } from "@takazudo/zudo-doc/doc-body-end";
-import { BodyEndIslands } from "./_body-end-islands";
-
-export const DocBodyEnd = createDocBodyEnd({ settings, BodyEndIslands });
