--- pages/lib/_math-block.tsx
@@ -1,4 +1,0 @@
-// Re-export from the shared package — moved to @takazudo/zudo-doc/math-block
-// as part of the package-first migration (epic #2321, S4 #2327).
-export type { MathBlockProps } from "@takazudo/zudo-doc/math-block";
-export { MathBlock } from "@takazudo/zudo-doc/math-block";
