+++ pages/lib/_nav-source-docs.ts
@@ -1,37 +1,17 @@
-// Thin stub — nav-source-docs moved to the package (epic #2344, S6).
-// Calls `createNavSourceDocs(ctx)` from @takazudo/zudo-doc/nav-source-docs
-// with the host singletons injected, then re-exports the resulting resolver
-// functions so all existing call sites continue to work unchanged.
+// Thin shim — nav-source resolvers ride on the unified ChromeContext now
+// (epic Collapse Wiring Shells #2420, FACTORIES #2424). The reconstructed
+// `createRouteContext` (in _chrome-context.ts) builds the identity-stable
+// nav-source API as part of the route context; this module just re-exports the
+// bindings so the existing call sites (route files, nav wrappers, _nav-data-prep,
+// route-enumerators) keep importing them from here unchanged.
 
-import { createNavSourceDocs } from "@takazudo/zudo-doc/nav-source-docs";
 export type { NavSourceDocs, NavSourceOptions } from "@takazudo/zudo-doc/nav-source-docs";
-import { defaultLocale, getLocaleConfig } from "@/config/i18n";
-import { settings } from "@/config/settings";
-import { loadCategoryMeta, isNavVisible } from "@/utils/docs";
-import { isDefaultLocaleOnlyPath } from "@/utils/base";
-import { stableDocs } from "./_nav-source-cache";
+import { routeContext } from "./_route-context";
 
-const {
+export const {
   resolveNavSource,
   resolveVersionedLocaleSource,
   loadNavSourceDocs,
   stableMergeCategoryMeta,
   stableNavDocs,
-} = createNavSourceDocs({
-  defaultLocale,
-  docsDir: settings.docsDir,
-  getVersions: () => settings.versions,
-  getLocaleConfig,
-  loadCategoryMeta,
-  isNavVisible,
-  isDefaultLocaleOnlyPath,
-  stableDocs,
-});
-
-export {
-  resolveNavSource,
-  resolveVersionedLocaleSource,
-  loadNavSourceDocs,
-  stableMergeCategoryMeta,
-  stableNavDocs,
-};
+} = routeContext;
