--- pages/lib/_search-widget-script.ts
@@ -1,2 +1,0 @@
-// Host thin-stub — see @takazudo/zudo-doc/search-widget-script (epic #2344, S5).
-export { SEARCH_WIDGET_SCRIPT } from "@takazudo/zudo-doc/search-widget-script";
