--- pages/lib/_toc-title.ts
@@ -1,3 +1,0 @@
-// TOC section-label resolver for the Toc/MobileToc overrides mounted by
-// `_doc-page-shell.tsx`, re-exported from the @takazudo/zudo-doc package.
-export { getTocTitle } from "@takazudo/zudo-doc/toc";
